27 [QA] added readme and renamed package
