+++ lagerta-jmh/README.md
@@ -1,0 +1,4 @@
+### JMH Module
+
+    1. mvn clean package -DskipTests -f lagerta-core
+    2. mvn clean test -f lagerta-jmh -P benchmark
